requirements.txt
fastapi
uvicorn[standard]
biopython
python-dotenv